pyproject.toml
[project]
name = "streaming-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.0.0",
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "dishka>=1.7.2",
    "fastapi>=0.128.0",
    "faststream[rabbit]>=0.6.5",
    "orjson>=3.11.5",
    "prometheus-client>=0.23.1",
    "psycopg[binary]>=3.3.2",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.45",
    "uuid7>=0.1.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
lint = [
    "ruff>=0.14.11",
]
test = [
    "locust>=2.43.0",
]

[build-system]
requires = ["hatchling>=1.13"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/streaming_service"]

[tool.ruff]
src = ["src"]
include = ["src/**.py", "tests/**.py"]
target-version = "py312"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]

[tool.ruff.lint.per-file-ignores]
"src/streaming_service/infrastructure/persistence/migrations/env.py" = ["ERA001", "INP001", "E402"]
"src/streaming_service/infrastructure/persistence/migrations/versions/**" = ["ALL"]
"tests/**" = ["S101"]

[tool.locust]
headless = true
locustfile = "scripts/locustfile.py"
host = "http://localhost:8000"
users = 100
spawn-rate = 30
run-time = "30s"
